src/main/java/nextstep/lms/domain/SessionDate.java
package nextstep.lms.domain;

import java.time.LocalDate;

public class SessionDate {

    private LocalDate startDate;
    private LocalDate endDate;

    public SessionDate(LocalDate startDate, LocalDate endDate) {
        validateDate(startDate, endDate);
        this.startDate = startDate;
        this.endDate = endDate;
    }

    private void validateDate(LocalDate startDate, LocalDate endDate) {
        if (endDate.isBefore(startDate)) {
            throw new IllegalArgumentException("종료 날짜가 시작 날짜보다 앞일 수 없습니다.");
        }
    }

    public boolean isAfterEndDate(LocalDate date) {
        return date.isAfter(endDate);
    }

    public LocalDate getStartDate() {
        return startDate;
    }

    public LocalDate getEndDate() {
        return endDate;
    }
}
